Allow Handlers to be exclusive too
As well as their results being exclusive we should allow handlers to be
exclusive too. This is because a handler might want to pre-empt the
calling of other handlers if it is going to have side effects or it
believes that it controls something very specific to itself.

For example, if I ask, "What is the time?" then a time service might want
to exclusively claim it so that, say, a Wikipedia service doesn't wander
off to fetch a definition of "time".