Don't adjust thresholds when someone's talking
If we think that someone is talking then we should not adjust the trigger
threshold, or accumulate into the "background" average. When they've stopped
talking then we're back to being ambient so we resume that accumulation. This
stops the problem of someone's speech pushing up the trigger threshold, causing
the recording to stop as a result.