Make all decoding asynchronous
Make the decoding and appending of tokens completely asynchronous so that we can
go back to listening for input right away. This makes the system more responsive.

As part of this we also do away with the logic to not listen when we're speaking
ourselves, since it gets in the way of this and wasn't that useful anyhow.