Added Unicorn HAT Mini
Added support for the Mini version of the Unicorn HAT. This was mostly just a
refactoring of the HD class so that it could work for both HAT types.

A side-effect of this change was to rename the original notifier to be
more specific about the type of HAT, it's now got an 'Hd' in it.