First cut of input support using Whisper
OpenAI open-sourced their Whisper STT engine. This requires a GPU to
work (and PyTorch) so it's not something which can be run on a Pi but we
look to add support for it. Part of this is done via a simple server
application, which can be accessed via the `remote` audio module.

This is the first cut of getting it to work in server mode. It also
seems to work in local mode too.