Make matching threshold configurable
It's possible that the matching threshold is too strict so allow users
to tune it down via a param.

Improve searching to attempt to search with the artist in order to
narrow it down a bit better.

And if we don't find a match then say so, since we might be the only
handler which matched and it's misleading to say that we don't know how
to handle the request.